gui-wm/labwc: Correct PATCHES to work with 9999
The PATCHES variable was added to the 9999 ebuild. Because there wasn't
a `labwc-9999-meson_doc_path.patch`, the live version of the ebuild was
failing.

I've renamed the PATCH file to `labwc-meson_doc_path.patch` so that
the 0.7.4 ebuild as well as the 9999 ebuild can make use of it. I've
also ammended the 0.7.4 and 9999 ebuild to make use of this renamed
file via `${PN}-meson_doc_path.patch`.

Running `pkgcheck scan` was complaining about incorrect spacing on the
PATCHES array so I also corrected that.

Signed-off-by: Daniella Kicsak <nosqrt@outlook.com>